requirements.txt
beautifulsoup4==4.5.1
cffi==1.5.2
cryptography==1.3.1
enum34==1.1.3
html2text==2016.9.19
idna==2.1
ipaddress==1.0.16
itsdangerous==0.24
MarkupSafe==0.23
ndg-httpsclient==0.4.0
py==1.4.31
pyasn1==0.1.9
pycparser==2.14
PyInstaller==3.2
pyOpenSSL==16.0.0
pytest==2.9.1
requests==2.9.1
six==1.10.0
Werkzeug==0.11.8
